Added in prevention of win condition, precluding completed chores

+++ source/core/src/main/com/deco2800/game/utils/math/Vector2Utils.java
@@ -45,7 +45,7 @@ public class Vector2Utils {
 
   public static Vector2 read(JsonValue jsonData) {
     int[] points = jsonData.asIntArray();
-    return new Vector2(points[0], points[1]);
+    return new Vector2((float) points[0], (float) points[1]);
   }
 
   private Vector2Utils() {
